pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aether-runner"
version = "0.1.0"
description = "Aether multimodal native fallback runner"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.3.0",
  "PyYAML>=6.0.1",
  "python-multipart>=0.0.9",
  "httpx>=0.27.0",
  "prometheus-client>=0.21.0",
  "numpy>=2.0.0",
  "Pillow>=10.4.0"
]

[project.optional-dependencies]
inference = [
  "torch>=2.4.0",
  "transformers>=4.57.0",
  "accelerate>=1.0.0"
]
test = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "httpx>=0.27.0"
]

[tool.setuptools.packages.find]
include = ["aether_runner*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
